cleaned up app and updated font checking
