+++ package.json
@@ -8,6 +8,7 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@expo-google-fonts/inter": "^0.1.0",
     "@expo/vector-icons": "^10.2.1",
     "@expo/webpack-config": "^0.12.40",
     "@react-native-community/masked-view": "0.1.10",
@@ -24,7 +25,6 @@
     "react-native": "^0.63.3",
     "react-native-easy-grid": "^0.2.2",
     "react-native-gesture-handler": "~1.7.0",
-    "react-native-google-signin": "^2.1.1",
     "react-native-reanimated": "~1.13.0",
     "react-native-safe-area-context": "3.1.4",
     "react-native-screens": "~2.10.1",
